requirements.txt
flask==3.0.0
flask-cors
requests
pyTelegramBotAPI==4.14.0
gunicorn==21.2.0
python-dotenv==1.0.0
python-telegram-bot==13.7
flask_socketio
eventlet
